+++ src/main/java/com/team/comma/controller/TestController.java
@@ -1,0 +1,38 @@
+package com.team.comma.controller;
+
+import com.team.comma.entity.TestEntity;
+import com.team.comma.repository.TestRepository;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.Parameter;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
+
+@RestController
+public class TestController {
+    @Autowired
+    TestRepository testrepos;
+
+    @RequestMapping(value = "/" , method = RequestMethod.GET)
+    public List<TestEntity> results() {
+        List<TestEntity> result = testrepos.findAll();
+
+        return result;
+    }
+
+    @Operation(summary = "test hello", description = "hello api example")
+    @ApiResponses({
+            @ApiResponse(responseCode = "200", description = "OK !!"),
+            @ApiResponse(responseCode = "400", description = "BAD REQUEST !!"),
+            @ApiResponse(responseCode = "404", description = "NOT FOUND !!"),
+            @ApiResponse(responseCode = "500", description = "INTERNAL SERVER ERROR !!")
+    })
+    @GetMapping("/hello")
+    public ResponseEntity<String> hello(@Parameter(description = "이름", required = true, example = "Park") @RequestParam String name) {
+        return ResponseEntity.ok("hello " + name);
+    }
+}
