+++ src/main/java/com/team/comma/repository/TestRepository.java
@@ -1,6 +1,6 @@
-package com.team.valueup.repository;
+package com.team.comma.repository;
 
-import com.team.valueup.entity.TestEntity;
+import com.team.comma.entity.TestEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TestRepository extends JpaRepository<TestEntity, String> {
