+++ src/main/resources/application.properties
@@ -1,6 +1,6 @@
 server.port=8080
 # LocalServer ??
-spring.datasource.url=jdbc:mysql://localhost:3306/valueup?useSSL=false&allowPublicKeyRetrieval=true&characterEncoding=UTF-8&serverTimezone=UTC
+spring.datasource.url=jdbc:mysql://localhost:3306/comma?useSSL=false&allowPublicKeyRetrieval=true&characterEncoding=UTF-8&serverTimezone=UTC
 # DockerServer ??
 # spring.datasource.url=jdbc:mysql://todayCheckMysql:3306/todaycheck?useSSL=false&allowPublicKeyRetrieval=true&characterEncoding=UTF-8&serverTimezone=UTC
 spring.datasource.username=root
